Move resource provider help fucntions to ResourceProviderHelper class
