+++ src/PSRule.Rules.Azure/Runtime/Helper.cs
@@ -8,11 +8,6 @@ using PSRule.Rules.Azure.Data.Template;
 using PSRule.Rules.Azure.Pipeline;
 using PSRule.Rules.Azure.Pipeline.Output;
 using System.Management.Automation;
-using System.Collections.Generic;
-using System.IO;
-using Newtonsoft.Json;
-using System.Linq;
-using System;
 
 namespace PSRule.Rules.Azure.Runtime
 {
@@ -21,10 +16,6 @@ namespace PSRule.Rules.Azure.Runtime
     /// </summary>
     public static class Helper
     {
-        private readonly static string AssemblyPath = Path.GetDirectoryName(typeof(Helper).Assembly.Location);
-        private const string DATAFILE_PROVIDERS = "providers.json";
-        private static Dictionary<string, ResourceProvider> _Providers;
-
         public static string CompressExpression(string expression)
         {
             if (!IsTemplateExpression(expression))
@@ -69,6 +60,12 @@ namespace PSRule.Rules.Azure.Runtime
             return builder;
         }
 
+        public static ResourceProviderType[] GetResourceType(string providerNamespace, string resourceType)
+        {
+            var resourceProviderHelper = new ResourceProviderHelper();
+            return resourceProviderHelper.GetResourceType(providerNamespace, resourceType);
+        }
+
         #region Helper methods
 
         private static PipelineContext GetContext(PSCmdlet commandRuntime = null)
@@ -76,41 +73,6 @@ namespace PSRule.Rules.Azure.Runtime
             var option = PSRuleOption.FromFileOrDefault(PSRuleOption.GetWorkingPath());
             var context = new PipelineContext(option, commandRuntime != null ? new PSPipelineWriter(option, commandRuntime) : null);
             return context;
-        }
-
-        private static Dictionary<string, T> ReadDataFile<T>(string fileName)
-        {
-            var sourcePath = Path.Combine(AssemblyPath, fileName);
-            if (!File.Exists(sourcePath))
-                return null;
-
-            var json = File.ReadAllText(sourcePath);
-            var settings = new JsonSerializerSettings();
-            settings.Converters.Add(new ResourceProviderConverter());
-            var result = JsonConvert.DeserializeObject<Dictionary<string, T>>(json, settings);
-            return result;
-        }
-
-        private static Dictionary<string, ResourceProvider> ReadProviders()
-        {
-            return ReadDataFile<ResourceProvider>(DATAFILE_PROVIDERS);
-        }
-
-        public static ResourceProviderType[] GetResourceType(string providerNamespace, string resourceType)
-        {
-            if (_Providers == null)
-                _Providers = ReadProviders();
-
-            if (_Providers == null || _Providers.Count == 0 || !_Providers.TryGetValue(providerNamespace, out ResourceProvider provider))
-                return Array.Empty<ResourceProviderType>();
-
-            if (resourceType == null)
-                return provider.Types.Values.ToArray();
-
-            if (!provider.Types.ContainsKey(resourceType))
-                return Array.Empty<ResourceProviderType>();
-
-            return new ResourceProviderType[] { provider.Types[resourceType] };
         }
 
         #endregion Helper methods
